+++ .vscode/settings.json
@@ -18,9 +18,44 @@
     ],
     "version": 4,
     "C_Cpp.default.configurationProvider": "mesonbuild.mesonbuild",
+    "C_Cpp.codeAnalysis.runAutomatically": true,
+    "C_Cpp.errorSquiggles": "enabled",
+    "C_Cpp.intelliSenseEngine": "default",                                               // Use Tag Parser for better performance with large projects
+    "C_Cpp.intelliSenseEngineFallback": "Disabled",                                         // Disable fallback to avoid performance issues
+    "C_Cpp.default.includePath": [
+        "${workspaceFolder}/include",
+        "${workspaceFolder}/src",
+        "${workspaceFolder}/build/meson-private",
+        "${workspaceFolder}/build/meson-uninstalled"
+    ],
+    "C_Cpp.default.defines": [
+        "HAVE_CONFIG_H",                                                                    // Define if you have a config.h generated by Meson
+        "MESON_BUILD"                                                                       // Define to indicate this is a Meson build
+    ],
+    "C_Cpp.default.browse.path": [
+        "${workspaceFolder}/include",
+        "${workspaceFolder}/src",
+        "${workspaceFolder}/build/meson-private",
+        "${workspaceFolder}/build/meson-uninstalled"
+    ],
+    "C_Cpp.default.browse.limitSymbolsToIncludedHeaders": true,                             // Limit symbols to included headers for performance
+    "C_Cpp.default.browse.databaseFilename": "${workspaceFolder}/.vscode/browse.db",        // Path to the browse database
+    "C_Cpp.default.browse.autoDetect": true,                                                // Automatically detect browse paths
+    "C_Cpp.default.browse.useDefaultDatabase": false,                                       // Use a custom database file    
     "files.associations": {
         "*.c": "c",
         "common.h": "c",
-        "port.h": "c"
+        "port.h": "c",
+        "cstdlib": "c",
+        "headers.h": "c",
+        "rte_ether.h": "c",
+        "crypto.h": "c",
+        "rte_mbuf.h": "c",
+        "rte_mbuf_dyn.h": "c",
+        "rte_mempool.h": "c",
+        "string.h": "c",
+        "ctype.h": "c",
+        "transit.h": "c",
+        "ingress.h": "c"
     }
 }
