+++ src/core/init.c
@@ -1,0 +1,30 @@
+#include "core/init.h"
+
+#include <rte_mbuf_core.h>
+#include <rte_mbuf_dyn.h>
+#include <stdlib.h>
+
+#include "utils/common.h"
+
+void register_tsc_dynfield() {
+  static const struct rte_mbuf_dynfield tsc_dynfield_desc = {
+      .name = "dpdk_pot_dynfield_tsc",
+      .size = sizeof(tsc_t),
+      .align = alignof(tsc_t),
+  };
+  tsc_dynfield_offset = rte_mbuf_dynfield_register(&tsc_dynfield_desc);
+  if (tsc_dynfield_offset < 0) rte_exit(EXIT_FAILURE, "Cannot register mbuf field\n");
+}
+
+void init_eal(int argc, char *argv[]) {
+  int ret = rte_eal_init(argc, argv);
+  if (ret < 0) rte_exit(EXIT_FAILURE, "Error with EAL initialization\n");
+}
+
+struct rte_mempool *create_mempool() {
+  struct rte_mempool *mbuf_pool =
+      rte_pktmbuf_pool_create("MBUF_POOL", NUM_MBUFS * rte_eth_dev_count_avail(), MBUF_CACHE_SIZE, 0,
+                              RTE_MBUF_DEFAULT_BUF_SIZE + EXTRA_SPACE, rte_socket_id());
+  if (mbuf_pool == NULL) rte_exit(EXIT_FAILURE, "Cannot create mbuf pool\n");
+  return mbuf_pool;
+}
