+++ src/latency.c
@@ -1,4 +1,4 @@
-#include <include/latency.h>
+#include "latency.h"
 
 struct latency_numbers_t latency_numbers = {0, 0, 0};
