+++ src/port.c
@@ -1,6 +1,6 @@
 #include "port.h"
 
-#include "common.h"
+#include "utils/common.h"
 #include "latency.h"
 
 /*
